Improve walletUnspents tests
Not fully fixed but improved.

These tests rely on some preconditions that are not always met:
balance and unspent count.

If a precondition is not met, we can create transactions so that the
conditions are met when the transactions are confirmed and skip the
tests.

This is currently not implemented for wallet balance. We could fund
the wallets from a central faucet, but for now we will show the
deposit address in the error message.

If we have an insufficient number of unspents, we can always call
`fanout`. Since this was part of a test already, I decided to split
the tests and conditionally skip the test that requires a minimum
unspent count.

Other changes:

* Use `getWalletWithMinBalance()` wrapper to simplify balance checking
  and make funding simpler.
* Wrap `Promise.delay()` in custom function that logs what is
  happening.